fix CodeCheck

+++ frameworks/innerkitsimpl/converter/src/post_proc.cpp
@@ -198,19 +198,20 @@ bool PostProc::CopyPixels(PixelMap& pixelMap, uint8_t* dstPixels, const Size& ds
     uint8_t *srcPixels = const_cast<uint8_t *>(pixelMap.GetPixels()) + (top * srcWidth + left) * pixelBytes;
     uint8_t *dstStartPixel = nullptr;
     uint8_t *srcStartPixel = nullptr;
-    uint32_t targetRowBytes = targetWidth * pixelBytes;
+    int32_t targetRowBytes = targetWidth * pixelBytes;
     if (targetRowStride <= 0) {
         targetRowStride = targetRowBytes;
     }
-    uint32_t srcRowBytes = srcWidth * pixelBytes;
+    int32_t srcRowBytes = srcWidth * pixelBytes;
     if (srcRowStride <= 0) {
         srcRowStride = srcRowBytes;
     }
-    uint32_t copyRowBytes = std::min(srcWidth, targetWidth) * pixelBytes;
+    int32_t copyRowBytes = std::min(srcWidth, targetWidth) * pixelBytes;
     for (int32_t scanLine = 0; scanLine < std::min(srcHeight, targetHeight); scanLine++) {
         dstStartPixel = dstPixels + scanLine * targetRowStride;
         srcStartPixel = srcPixels + scanLine * srcRowStride;
-        errno_t errRet = memcpy_s(dstStartPixel, targetRowBytes, srcStartPixel, copyRowBytes);
+        errno_t errRet = memcpy_s(dstStartPixel, static_cast<size_t>(targetRowBytes),
+            srcStartPixel, static_cast<size_t>(copyRowBytes));
         if (errRet != EOK) {
             IMAGE_LOGE("[PostProc]memcpy scanline %{public}d fail, errorCode = %{public}d", scanLine, errRet);
             return false;
@@ -760,8 +761,8 @@ bool PostProc::ScalePixelMapEx(const Size &desiredSize, PixelMap &pixelMap, cons
         IMAGE_LOGE("pixelMap format is invalid, format: %{public}d", imgInfo.pixelFormat);
         return false;
     }
-    uint32_t dstBufferSize = desiredSize.height * desiredSize.width * ImageUtils::GetPixelBytes(imgInfo.pixelFormat);
-    MemoryData memoryData = {nullptr, dstBufferSize, "ScalePixelMapEx ImageData", desiredSize};
+    int64_t dstBufferSize = desiredSize.height * desiredSize.width * ImageUtils::GetPixelBytes(imgInfo.pixelFormat);
+    MemoryData memoryData = {nullptr, static_cast<size_t>(dstBufferSize), "ScalePixelMapEx ImageData", desiredSize};
     auto mem = MemoryManager::CreateMemory(pixelMap.GetAllocatorType(), memoryData);
     if (mem == nullptr) {
         IMAGE_LOGE("ScalePixelMapEx CreateMemory failed");
